refactor: streamline error handling and improve code clarity in IOCP driver
- Refactored error reporting in the `RioState` and `RioReportExt` implementations to enhance clarity and maintainability.
- Simplified the handling of error codes and diagnostics in the `last_wsa_report` method, ensuring consistent error recovery across RIO operations.
- Improved the management of socket states by utilizing the `is_iocp_fallback` property more effectively in various methods.
- Cleaned up code formatting and structure for better readability and organization throughout the affected modules.

+++ platform-drivers/veloq-driver-iocp/src/rio/runtime/pool.rs
@@ -136,7 +136,7 @@ pub(super) struct CompletionActions {
 impl UdpRecvPool {
     #[inline]
     pub(super) fn is_datagram_completion(res: &RIORESULT) -> bool {
-        (res.Status == 0 || res.Status == WSAEMSGSIZE as i32) && res.BytesTransferred > 0
+        (res.Status == 0 || res.Status == WSAEMSGSIZE) && res.BytesTransferred > 0
     }
 
     pub(crate) fn uninit() -> Self {
